Finish and clean uop
