+++ packages/app/src/domain/topical/components/search/index.ts
@@ -1,0 +1,1 @@
+export * from './search';
