+++ packages/app/src/utils/use-is-mounted.ts
@@ -1,0 +1,7 @@
+import { useEffect, useState } from 'react';
+
+export function useIsMounted() {
+  const [isMounted, setIsMounted] = useState(false);
+  useEffect(() => setIsMounted(true), []);
+  return isMounted;
+}
